Did some more refactoring and broke something on the client

+++ backend/src/staff/handlers/products.ts
@@ -9,13 +9,14 @@ import { Response, NextFunction } from "express";
 import prisma from "../../utils/db";
 import { AuthenticatedRequest, ProductCreateUpdateRequest } from "../../types";
 import { CATEGORIES } from "@prisma/client";
+import { serverError } from "../../utils/globals";
 
 // Function to fetch products
 export const fetchProducts = async (
   req: AuthenticatedRequest,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
     const page = req.query.page ? parseInt(req.query.page as string) : 1;
     const limit = req.query.limit ? parseInt(req.query.limit as string) : 10;
@@ -53,13 +54,17 @@ export const fetchProducts = async (
       products.pop();
     }
 
-    return res.status(200).json({
+    res.status(200).json({
       message: "Products fetched successfully",
       products,
       hasNext,
     });
   } catch (e) {
-    return next(e as Error);
+    if (e instanceof Error) {
+      next(e);
+      return;
+    }
+    next(serverError);
   }
 };
 
@@ -68,7 +73,7 @@ export const fetchIndividualProduct = async (
   req: AuthenticatedRequest,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
     const { id } = req.params;
 
@@ -87,18 +92,23 @@ export const fetchIndividualProduct = async (
     });
 
     if (!product) {
-      return res.status(404).json({
+      res.status(404).json({
         message: "Product not found",
         errorCode: "P400",
       });
+      return;
     }
 
-    return res.status(200).json({
+    res.status(200).json({
       message: "Product fetched successfully",
       product,
     });
   } catch (e) {
-    return next(e as Error);
+    if (e instanceof Error) {
+      next(e);
+      return;
+    }
+    next(serverError);
   }
 };
 
@@ -107,7 +117,7 @@ export const updateProduct = async (
   req: ProductCreateUpdateRequest,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
     const { id } = req.params;
     const { name, description, price, category, inventory } = req.body;
@@ -155,12 +165,16 @@ export const updateProduct = async (
       },
     );
 
-    return res.status(200).json({
+    res.status(200).json({
       message: "Product updated successfully",
       product,
     });
   } catch (e) {
-    return next(e as Error);
+    if (e instanceof Error) {
+      next(e);
+      return;
+    }
+    next(serverError);
   }
 };
 
@@ -169,20 +183,46 @@ export const deleteProduct = async (
   req: AuthenticatedRequest,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
     const { id } = req.params;
 
-    const product = await prisma.product.delete({
+    const product = await prisma.product.findUnique({
       where: {
         id,
       },
+      include: {
+        images: true,
+        seller: {
+          include: {
+            image: true,
+          },
+        },
+      },
     });
 
-    return res.status(203).json({
+    if (!product) {
+      res.status(404).json({
+        message: "Product not found",
+        errorCode: "P400",
+      });
+      return;
+    }
+
+    await prisma.product.delete({
+      where: {
+        id: product.id,
+      },
+    });
+
+    res.status(203).json({
       message: "Product deleted successfully",
     });
   } catch (e) {
-    return next(e as Error);
+    if (e instanceof Error) {
+      next(e);
+      return;
+    }
+    next(serverError);
   }
 };
